// src/generated/services/Fl_dev_cu_template_categoryService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { IOperationResult } from '@microsoft/power-apps/data';
import type { ManualTriggerInput } from '../models/Fl_dev_cu_template_categoryModel';
import type { ResponseActionOutput } from '../models/Fl_dev_cu_template_categoryModel';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';

export class Fl_dev_cu_template_categoryService {
  private static readonly dataSourceName = 'fl_dev_cu_template_category';

  private static readonly client = getClient(dataSourcesInfo);

  public static async Run(input: ManualTriggerInput): Promise<IOperationResult<ResponseActionOutput>> {
    const params: { input: ManualTriggerInput } = { input };
    const allParams = { ...params, "api-version": "2015-02-01-preview" };
    const result = await Fl_dev_cu_template_categoryService.client.executeAsync<{ input: ManualTriggerInput }, ResponseActionOutput>(
      {
        connectorOperation: {
          tableName: Fl_dev_cu_template_categoryService.dataSourceName,
          operationName: 'Run',
          parameters: allParams
        },
      });
    return result;
  }
}