// src/generated/models/Dev_tablecategorytemplatesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dev_tablecategorytemplatesdev_granularity = {
  775730001: 'General',
  775730002: 'Porcr_dito'
} as const;
export type Dev_tablecategorytemplatesdev_granularity = keyof typeof Dev_tablecategorytemplatesdev_granularity;
export const Dev_tablecategorytemplatesdev_scope = {
  775730000: 'GENERAL',
  775730001: 'CREDITO'
} as const;
export type Dev_tablecategorytemplatesdev_scope = keyof typeof Dev_tablecategorytemplatesdev_scope;
export const Dev_tablecategorytemplatesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dev_tablecategorytemplatesstatecode = keyof typeof Dev_tablecategorytemplatesstatecode;
export const Dev_tablecategorytemplatesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dev_tablecategorytemplatesstatuscode = keyof typeof Dev_tablecategorytemplatesstatuscode;

export interface Dev_tablecategorytemplatesBase {
  dev_description?: string;
  dev_granularity: Dev_tablecategorytemplatesdev_granularity;
  dev_isrequired: boolean;
  dev_namecategory: string;
  dev_namecategorynormalized: string;
  dev_scope?: Dev_tablecategorytemplatesdev_scope;
  dev_tablecategorytemplateid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dev_tablecategorytemplatesstatecode;
  statuscode?: Dev_tablecategorytemplatesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dev_tablecategorytemplates extends Dev_tablecategorytemplatesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dev_granularityname?: string;
  dev_isrequiredname?: string;
  dev_scopename?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}